Refactor main.cpp and RenderSystem.cpp
This commit refactors the main.cpp file by uncommenting the code to add a MusicComponent and a SoundComponent to the ComponentManager. It also removes a call to the hello() function in the main loop. In RenderSystem.cpp, the commit adds event handling code to close the window when the user clicks the close button.