[UPD] app && interface && CMakeLists.txt
in the interface remove some file that are not needed any more, and change some other to refacto the server to create a server system. In the app folder adapt the files to the new server system. In the CMakeLists.txt add the verification of the user's distro and install the SFML